refactor(search): derive validation rules from the kind-to-operator table
- move OPERATOR_BY_KIND and filterOperatorFor into schema.ts, next to FieldKind:
  kind semantics live with the kinds; query.ts re-exports them unchanged
- validateSearchType derives its rules from the table instead of restating them:
  a kind without a where operator is neither filterable nor facetable, and only
  range-operator kinds may declare facetRanges — so validation, the surfaces and
  the compilers share one source and cannot drift
